+++ veltahleon/src/main/java/com/testament/veltahleon/rest/WorldController.java
@@ -1,0 +1,31 @@
+package com.testament.veltahleon.rest;
+
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+@RestController
+@RequestMapping("/world_map")
+public class WorldController {
+
+    @GetMapping(value = "/pure_landmass/image/{fileName}", produces = {MediaType.IMAGE_PNG_VALUE, MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_GIF_VALUE})
+    public byte[] getWorldMapPureLandmass(@PathVariable("fileName") String fileName) throws IOException {
+        return Files.readAllBytes(Paths.get("src/main/java/com/testament/veltahleon/assets/images/world_maps/" + fileName));
+    }
+
+    @GetMapping(value = "/heightmap/image/{fileName}", produces = {MediaType.IMAGE_PNG_VALUE, MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_GIF_VALUE})
+    public byte[] getWorldMapHeightMap(@PathVariable("fileName") String fileName) throws IOException {
+        return Files.readAllBytes(Paths.get("src/main/java/com/testament/veltahleon/assets/images/world_maps/" + fileName));
+    }
+
+    @GetMapping(value = "/nations/image/{fileName}", produces = {MediaType.IMAGE_PNG_VALUE, MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_GIF_VALUE})
+    public byte[] getWorldMapNations(@PathVariable("fileName") String fileName) throws IOException {
+        return Files.readAllBytes(Paths.get("src/main/java/com/testament/veltahleon/assets/images/world_maps/" + fileName));
+    }
+}