+++ veltahleon/src/main/java/com/testament/veltahleon/services/implementation/places/NationServiceImpl.java
@@ -88,6 +88,10 @@ public class NationServiceImpl implements NationService {
             newNation.setType(nation.getType());
         }
 
+        if(nation.getUrlFlag() != null && newNation.getUrlFlag() != nation.getUrlFlag()) {
+            newNation.setUrlFlag(nation.getUrlFlag());
+        }
+
         return nationRepository.save(newNation);
     }
 }